AWS CLI v2 resolves pulling containers from s3... (#14)
along with some other issues. We had updated the AWS CLI version within the base-container,
however the scripts that run commands to pull from S3 are still leveraging and attempting
to install AWS CLI v1. This should be the last place we are using the AWS CLI v1. This also
resolves thhe current issue of not being able to spin up new starphleet ships.